third_party/WebKit/LayoutTests/fast/dom/MutationObserver/script-end-tag-expected.txt

Test that a script end tags invokes MutationObserver callbacks

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mutationsDelivered is false
PASS mutationsDelivered is true
PASS successfullyParsed is true

TEST COMPLETE
